test: . Move functions for creating statuses to TestingTools/StatusExamples.ts
