+++ tests/DocumentationSamples/DocsSamplesForStatuses.test.ts
@@ -10,6 +10,14 @@ import {
     verifyStatusesInMultipleFormats,
     verifyTransitionsAsMarkdownTable,
 } from '../TestingTools/VerifyStatuses';
+import {
+    doneTogglesToCancelled,
+    doneTogglesToCancelledWithUnconventionalSymbols,
+    importantCycle,
+    proCon,
+    todoToInProgressToDone,
+    variousNonTaskStatuses,
+} from '../TestingTools/StatusExamples';
 
 function constructStatuses(importedStatuses: StatusCollection) {
     const statuses: Status[] = [];
@@ -17,62 +25,6 @@ function constructStatuses(importedStatuses: StatusCollection) {
         statuses.push(Status.createFromImportedValue(importedStatus));
     });
     return statuses;
-}
-
-function doneTogglesToCancelled() {
-    const statuses: StatusCollection = [
-        [' ', 'Todo', '/', 'TODO'],
-        ['x', 'Done', '-', 'DONE'],
-        ['/', 'In Progress', 'x', 'IN_PROGRESS'],
-        ['-', 'Cancelled', ' ', 'CANCELLED'],
-    ];
-    return statuses;
-}
-
-function doneTogglesToCancelledWithUnconventionalSymbols() {
-    const statuses: StatusCollection = [
-        [' ', 'Todo', '*', 'TODO'],
-        ['*', 'Done', 'x', 'DONE'],
-        ['x', 'Cancelled', ' ', 'CANCELLED'],
-    ];
-    return statuses;
-}
-
-function variousNonTaskStatuses() {
-    const importantCycle: StatusCollection = [
-        ['b', 'Bookmark', 'b', 'NON_TASK'],
-        ['E', 'Example', 'E', 'NON_TASK'],
-        ['I', 'Information', 'I', 'NON_TASK'],
-        ['P', 'Paraphrase', 'P', 'NON_TASK'],
-        ['Q', 'Quote', 'Q', 'NON_TASK'],
-    ];
-    return importantCycle;
-}
-
-function importantCycle() {
-    const importantCycle: StatusCollection = [
-        ['!', 'Important', 'D', 'TODO'],
-        ['D', 'Doing - Important', 'X', 'IN_PROGRESS'],
-        ['X', 'Done - Important', '!', 'DONE'],
-    ];
-    return importantCycle;
-}
-
-function todoToInProgressToDone() {
-    const importantCycle: StatusCollection = [
-        [' ', 'Todo', '/', 'TODO'],
-        ['/', 'In Progress', 'x', 'IN_PROGRESS'],
-        ['x', 'Done', ' ', 'DONE'],
-    ];
-    return importantCycle;
-}
-
-function proCon() {
-    const importantCycle: StatusCollection = [
-        ['P', 'Pro', 'C', 'NON_TASK'],
-        ['C', 'Con', 'P', 'NON_TASK'],
-    ];
-    return importantCycle;
 }
 
 describe('DefaultStatuses', () => {
